feat: merge duplicate test and update regexp for matching max steps
