+++ public/hooks/use_investigation.test.tsx
@@ -5,32 +5,42 @@
 
 import React from 'react';
 import { renderHook, act } from '@testing-library/react-hooks';
-import { BehaviorSubject } from 'rxjs';
+import { of } from 'rxjs';
 import { useInvestigation } from './use_investigation';
 import { NotebookReactContext } from '../components/notebooks/context_provider/context_provider';
-import { InvestigationPhase } from '../../common/state/notebook_state';
-import { OpenSearchDashboardsContextProvider } from '../../../../src/plugins/opensearch_dashboards_react/public';
+import { InvestigationPhase, NotebookState } from '../../common/state/notebook_state';
+import { TopContextState } from '../../common/state/top_context_state';
+import { useOpenSearchDashboards } from '../../../../src/plugins/opensearch_dashboards_react/public';
+import * as mlCommonsApis from '../utils/ml_commons_apis';
+import { useNotebook } from './use_notebook';
+import { useToast } from './use_toast';
+import { isValidPERAgentInvestigationResponse } from '../../common/utils/per_agent';
+import { SharedMessagePollingService } from '../components/notebooks/components/hypothesis/investigation/services/shared_message_polling_service';
 
-// Mock dependencies
+jest.mock('../../../../src/plugins/opensearch_dashboards_react/public', () => ({
+  useOpenSearchDashboards: jest.fn(),
+  OpenSearchDashboardsContextProvider: ({ children }: any) => children,
+}));
+jest.mock('../utils/ml_commons_apis');
+jest.mock('./use_notebook');
+jest.mock('./use_toast');
+jest.mock(
+  '../components/notebooks/components/hypothesis/investigation/services/shared_message_polling_service'
+);
+jest.mock('../../common/utils/per_agent');
 jest.mock('react-use', () => ({
   useObservable: jest.fn((subject$, defaultValue) => {
+    if (subject$ && typeof subject$.subscribe === 'function') {
+      let value = defaultValue;
+      subject$.subscribe((v: any) => {
+        value = v;
+      });
+      return value;
+    }
     if (subject$ && typeof subject$.getValue === 'function') {
       return subject$.getValue();
     }
     return defaultValue;
-  }),
-}));
-
-jest.mock('./use_notebook', () => ({
-  useNotebook: () => ({
-    updateHypotheses: jest.fn(),
-    updateNotebookContext: jest.fn(),
-  }),
-}));
-
-jest.mock('./use_toast', () => ({
-  useToast: () => ({
-    addError: jest.fn(),
   }),
 }));
 
@@ -64,116 +74,70 @@ const createMockServices = () => ({
   },
 });
 
-const createMockNotebookContext = (initialState = {}) => {
-  const defaultState = {
-    id: 'test-notebook-123',
-    hypotheses: [],
-    context: {
-      value: {
-        dataSourceId: 'test-datasource',
-        initialGoal: 'Test question',
-      },
-    },
-    runningMemory: undefined,
-    historyMemory: undefined,
-    investigationError: undefined,
-    isNotebookReadonly: false,
-    currentUser: 'test-user',
-    investigationPhase: undefined,
-    title: 'Test Notebook',
-    path: 'test-path',
-    ...initialState,
-  };
+describe('useInvestigation', () => {
+  let mockServices: any;
+  let mockNotebookState: NotebookState;
+  let mockParagraphHooks: any;
+  let mockAddError: jest.Mock;
+  let mockUpdateHypotheses: jest.Mock;
+  let mockUpdateNotebookContext: jest.Mock;
 
-  const state$ = new BehaviorSubject(defaultState);
-  const paragraphStates$ = new BehaviorSubject([]);
+  beforeEach(() => {
+    jest.clearAllMocks();
 
-  return {
-    state: {
-      getValue$: () => state$,
-      getParagraphStates$: () => paragraphStates$,
-      getParagraphsValue: jest.fn(() => []),
-      value: defaultState,
-      updateValue: jest.fn((updates) => {
-        const newValue = { ...state$.getValue(), ...updates };
-        state$.next(newValue);
-      }),
-    },
-    paragraphHooks: {
+    mockServices = createMockServices();
+    mockAddError = jest.fn();
+    mockUpdateHypotheses = jest.fn().mockResolvedValue(undefined);
+    mockUpdateNotebookContext = jest.fn().mockResolvedValue(undefined);
+
+    (useToast as jest.Mock).mockReturnValue({ addError: mockAddError });
+    (useNotebook as jest.Mock).mockReturnValue({
+      updateHypotheses: mockUpdateHypotheses,
+      updateNotebookContext: mockUpdateNotebookContext,
+    });
+
+    (useOpenSearchDashboards as jest.Mock).mockReturnValue({
+      services: mockServices,
+    });
+
+    mockNotebookState = new NotebookState({
+      paragraphs: [],
+      id: 'test-notebook',
+      title: 'Test Notebook',
+      context: new TopContextState({ dataSourceId: 'ds-123' }),
+      dataSourceEnabled: false,
+      dateCreated: '',
+      dateModified: '',
+      isLoading: false,
+      path: '',
+      vizPrefix: '',
+      isNotebookReadonly: false,
+      topologies: [],
+    });
+
+    mockParagraphHooks = {
       createParagraph: jest.fn(),
       batchCreateParagraphs: jest.fn(),
       batchRunParagraphs: jest.fn(),
       runParagraph: jest.fn(),
       batchDeleteParagraphs: jest.fn(),
-    },
-  };
-};
-
-describe('useInvestigation Hook', () => {
-  let mockServices: any;
-
-  beforeEach(() => {
-    mockServices = createMockServices();
-    jest.clearAllMocks();
+    };
   });
 
-  const renderUseInvestigation = (contextOverrides = {}) => {
-    const context = createMockNotebookContext(contextOverrides);
-    const hookResult = renderHook(() => useInvestigation(), {
-      wrapper: ({ children }) => (
-        <OpenSearchDashboardsContextProvider services={mockServices}>
-          <NotebookReactContext.Provider value={context}>{children}</NotebookReactContext.Provider>
-        </OpenSearchDashboardsContextProvider>
-      ),
-    });
-    return { ...hookResult, context };
-  };
-
-  describe('isInvestigating derived state', () => {
-    it('should return true when investigationPhase is PLANNING', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: InvestigationPhase.PLANNING,
-      });
-
-      expect(result.current.isInvestigating).toBe(true);
-    });
-
-    it('should return true when investigationPhase is RETRIEVING_CONTEXT', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: InvestigationPhase.RETRIEVING_CONTEXT,
-      });
-
-      expect(result.current.isInvestigating).toBe(true);
-    });
-
-    it('should return true when investigationPhase is GATHERING_DATA', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: InvestigationPhase.GATHERING_DATA,
-      });
-
-      expect(result.current.isInvestigating).toBe(true);
-    });
-
-    it('should return false when investigationPhase is COMPLETED', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: InvestigationPhase.COMPLETED,
-      });
-
-      expect(result.current.isInvestigating).toBe(false);
-    });
-
-    it('should return false when investigationPhase is undefined', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: undefined,
-      });
-
-      expect(result.current.isInvestigating).toBe(false);
-    });
-  });
+  const wrapper = ({ children }: { children: React.ReactNode }) => (
+    <NotebookReactContext.Provider
+      value={{
+        state: mockNotebookState,
+        paragraphHooks: mockParagraphHooks,
+      }}
+    >
+      {children}
+    </NotebookReactContext.Provider>
+  );
 
   describe('Hook API', () => {
     it('should return all required functions', () => {
-      const { result } = renderUseInvestigation();
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
 
       expect(result.current).toHaveProperty('isInvestigating');
       expect(result.current).toHaveProperty('doInvestigate');
@@ -190,13 +154,76 @@ describe('useInvestigation Hook', () => {
     });
   });
 
+  describe('isInvestigating state', () => {
+    it('should return true for active investigation phases', () => {
+      const phases = [
+        { phase: InvestigationPhase.PLANNING, expected: true },
+        { phase: InvestigationPhase.RETRIEVING_CONTEXT, expected: true },
+        { phase: InvestigationPhase.GATHERING_DATA, expected: true },
+      ];
+
+      phases.forEach(({ phase, expected }) => {
+        const state = new NotebookState({
+          ...mockNotebookState.value,
+          investigationPhase: phase,
+        });
+        const hooks = mockParagraphHooks;
+        const { result } = renderHook(() => useInvestigation(), {
+          wrapper: ({ children }) => (
+            <NotebookReactContext.Provider value={{ state, paragraphHooks: hooks }}>
+              {children}
+            </NotebookReactContext.Provider>
+          ),
+        });
+        expect(result.current.isInvestigating).toBe(expected);
+      });
+    });
+
+    it('should return false for completed or undefined phases', () => {
+      const phases = [
+        { phase: InvestigationPhase.COMPLETED, expected: false },
+        { phase: undefined, expected: false },
+      ];
+
+      phases.forEach(({ phase, expected }) => {
+        const state = new NotebookState({
+          ...mockNotebookState.value,
+          investigationPhase: phase,
+        });
+        const hooks = mockParagraphHooks;
+        const { result } = renderHook(() => useInvestigation(), {
+          wrapper: ({ children }) => (
+            <NotebookReactContext.Provider value={{ state, paragraphHooks: hooks }}>
+              {children}
+            </NotebookReactContext.Provider>
+          ),
+        });
+        expect(result.current.isInvestigating).toBe(expected);
+      });
+    });
+
+    it('should set isInvestigating to false on error', async () => {
+      (mlCommonsApis.getMLCommonsConfig as jest.Mock).mockRejectedValue(new Error('API Error'));
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(result.current.isInvestigating).toBe(false);
+    });
+  });
+
   describe('checkOngoingInvestigation', () => {
     it('should return false when no ongoing investigation', async () => {
       mockHttp.get.mockResolvedValueOnce({
         runningMemory: undefined,
       });
 
-      const { result } = renderUseInvestigation();
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
 
       let hasOngoing;
       await act(async () => {
@@ -217,7 +244,7 @@ describe('useInvestigation Hook', () => {
         },
       });
 
-      const { result } = renderUseInvestigation();
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
 
       let hasOngoing;
       await act(async () => {
@@ -234,7 +261,7 @@ describe('useInvestigation Hook', () => {
     it('should return true on error', async () => {
       mockHttp.get.mockRejectedValueOnce(new Error('Network error'));
 
-      const { result } = renderUseInvestigation();
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
 
       let hasOngoing;
       await act(async () => {
@@ -251,14 +278,15 @@ describe('useInvestigation Hook', () => {
         value: { id: 'para-123' },
       };
 
-      const { result, context } = renderUseInvestigation();
-      context.paragraphHooks.createParagraph.mockResolvedValueOnce(mockParagraph);
+      mockParagraphHooks.createParagraph.mockResolvedValueOnce(mockParagraph);
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
 
       await act(async () => {
         await result.current.addNewFinding({ text: 'Test finding' });
       });
 
-      expect(context.paragraphHooks.createParagraph).toHaveBeenCalledWith({
+      expect(mockParagraphHooks.createParagraph).toHaveBeenCalledWith({
         index: expect.any(Number),
         input: {
           inputText: 'Test finding',
@@ -270,92 +298,384 @@ describe('useInvestigation Hook', () => {
         aiGenerated: false,
       });
 
-      expect(context.paragraphHooks.runParagraph).toHaveBeenCalledWith({
+      expect(mockParagraphHooks.runParagraph).toHaveBeenCalledWith({
         id: 'para-123',
       });
     });
 
     it('should not run paragraph if creation fails', async () => {
-      const { result, context } = renderUseInvestigation();
-      context.paragraphHooks.createParagraph.mockResolvedValueOnce(null);
+      mockParagraphHooks.createParagraph.mockResolvedValueOnce(null);
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
 
       await act(async () => {
         await result.current.addNewFinding({ text: 'Test finding' });
       });
 
-      expect(context.paragraphHooks.runParagraph).not.toHaveBeenCalled();
+      expect(mockParagraphHooks.runParagraph).not.toHaveBeenCalled();
     });
   });
 
-  describe('Phase transitions', () => {
-    it('should have undefined phase initially', () => {
-      const { result } = renderUseInvestigation();
-      expect(result.current.isInvestigating).toBe(false);
-    });
+  describe('executeInvestigation - executorMemoryId retry logic', () => {
+    const setupMocks = () => {
+      (mlCommonsApis.getMLCommonsConfig as jest.Mock).mockResolvedValue({
+        configuration: { agent_id: 'test-agent-id' },
+      });
+      (mlCommonsApis.getMLCommonsAgentDetail as jest.Mock).mockResolvedValue({
+        memory: { memory_container_id: 'test-container-id' },
+      });
+      (mlCommonsApis.executeMLCommonsAgent as jest.Mock).mockResolvedValue({
+        inference_results: [{ output: [{ result: 'interaction_id:test-interaction' }] }],
+      });
+    };
 
-    it('should transition through phases during investigation lifecycle', () => {
-      // This test verifies the hook properly derives isInvestigating from different phases
-      const phases = [
-        { phase: InvestigationPhase.PLANNING, expected: true },
-        { phase: InvestigationPhase.RETRIEVING_CONTEXT, expected: true },
-        { phase: InvestigationPhase.GATHERING_DATA, expected: true },
-        { phase: InvestigationPhase.COMPLETED, expected: false },
-      ];
+    it('should accept valid executorMemoryId on first attempt', async () => {
+      setupMocks();
+      (mlCommonsApis.createAgenticExecutionMemory as jest.Mock).mockResolvedValue({
+        session_id: 'valid-session-id',
+      });
 
-      phases.forEach(({ phase, expected }) => {
-        const { result } = renderUseInvestigation({
-          investigationPhase: phase,
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
         });
-
-        expect(result.current.isInvestigating).toBe(expected);
       });
+
+      expect(mlCommonsApis.createAgenticExecutionMemory).toHaveBeenCalledTimes(1);
+    });
+
+    it('should retry when executorMemoryId starts with - or _', async () => {
+      setupMocks();
+      (mlCommonsApis.createAgenticExecutionMemory as jest.Mock)
+        .mockResolvedValueOnce({ session_id: '-invalid-id' })
+        .mockResolvedValueOnce({ session_id: 'valid-session-id' });
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mlCommonsApis.createAgenticExecutionMemory).toHaveBeenCalledTimes(2);
+    });
+
+    it('should retry up to 3 times for invalid executorMemoryId', async () => {
+      setupMocks();
+      (mlCommonsApis.createAgenticExecutionMemory as jest.Mock)
+        .mockResolvedValueOnce({ session_id: '-invalid-1' })
+        .mockResolvedValueOnce({ session_id: '_invalid-2' })
+        .mockResolvedValueOnce({ session_id: 'valid-id-3' });
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mlCommonsApis.createAgenticExecutionMemory).toHaveBeenCalledTimes(3);
+    });
+
+    it('should throw error when executorMemoryId is null after retries', async () => {
+      setupMocks();
+      (mlCommonsApis.createAgenticExecutionMemory as jest.Mock)
+        .mockResolvedValueOnce({ session_id: null })
+        .mockResolvedValueOnce({ session_id: null })
+        .mockResolvedValueOnce({ session_id: null });
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mockAddError).toHaveBeenCalledWith(
+        expect.objectContaining({
+          title: 'Failed to execute per agent',
+        })
+      );
+    });
+
+    it('should stop retrying when valid id is found', async () => {
+      setupMocks();
+      (mlCommonsApis.createAgenticExecutionMemory as jest.Mock)
+        .mockResolvedValueOnce({ session_id: '-invalid' })
+        .mockResolvedValueOnce({ session_id: 'abc123' });
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mlCommonsApis.createAgenticExecutionMemory).toHaveBeenCalledTimes(2);
+      expect(mlCommonsApis.executeMLCommonsAgent).toHaveBeenCalledWith(
+        expect.objectContaining({
+          parameters: expect.objectContaining({
+            executor_agent_memory_id: 'abc123',
+          }),
+        })
+      );
+    });
+
+    it('should continue with invalid id after 3 retries all start with - or _', async () => {
+      setupMocks();
+      (mlCommonsApis.createAgenticExecutionMemory as jest.Mock)
+        .mockResolvedValueOnce({ session_id: '-invalid-1' })
+        .mockResolvedValueOnce({ session_id: '_invalid-2' })
+        .mockResolvedValueOnce({ session_id: '-invalid-3' });
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mlCommonsApis.createAgenticExecutionMemory).toHaveBeenCalledTimes(3);
+      expect(mlCommonsApis.executeMLCommonsAgent).toHaveBeenCalledWith(
+        expect.objectContaining({
+          parameters: expect.objectContaining({
+            executor_agent_memory_id: '-invalid-3',
+          }),
+        })
+      );
     });
   });
 
-  describe('Error states', () => {
-    it('should not be investigating when there is an error', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: InvestigationPhase.COMPLETED,
-        investigationError: 'Test error',
+  describe('executeInvestigation - error handling', () => {
+    it('should handle readonly notebook', async () => {
+      mockNotebookState.updateValue({ isNotebookReadonly: true });
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
       });
 
-      expect(result.current.isInvestigating).toBe(false);
+      expect(mockAddError).toHaveBeenCalledWith(
+        expect.objectContaining({
+          title: 'Failed to execute per agent',
+        })
+      );
     });
 
-    it('should still show investigating status during error in active phase', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: InvestigationPhase.GATHERING_DATA,
-        investigationError: 'Partial error',
+    it('should handle missing agentId', async () => {
+      (mlCommonsApis.getMLCommonsConfig as jest.Mock).mockResolvedValue({
+        configuration: { agent_id: null },
       });
 
-      expect(result.current.isInvestigating).toBe(true);
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mockAddError).toHaveBeenCalled();
     });
   });
 
-  describe('Running memory states', () => {
-    it('should recognize investigating state with running memory', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: InvestigationPhase.GATHERING_DATA,
-        runningMemory: {
-          executorMemoryId: 'mem-123',
-          parentInteractionId: 'parent-123',
+  describe('response validation error handling', () => {
+    let mockIsValidPERAgentInvestigationResponse: jest.Mock;
+    let mockSharedMessagePollingService: any;
+
+    beforeEach(() => {
+      mockIsValidPERAgentInvestigationResponse = (isValidPERAgentInvestigationResponse as unknown) as jest.Mock;
+
+      mockSharedMessagePollingService = {
+        poll: jest.fn(),
+      };
+
+      (SharedMessagePollingService.getInstance as jest.Mock) = jest
+        .fn()
+        .mockReturnValue(mockSharedMessagePollingService);
+
+      (mlCommonsApis.getMLCommonsConfig as jest.Mock).mockResolvedValue({
+        configuration: { agent_id: 'test-agent-id' },
+      });
+      (mlCommonsApis.getMLCommonsAgentDetail as jest.Mock).mockResolvedValue({
+        memory: { memory_container_id: 'test-container-id' },
+      });
+      (mlCommonsApis.createAgenticExecutionMemory as jest.Mock).mockResolvedValue({
+        session_id: 'test-session-id',
+      });
+      (mlCommonsApis.executeMLCommonsAgent as jest.Mock).mockResolvedValue({
+        response: {
+          parent_interaction_id: 'test-parent-interaction',
         },
       });
-
-      expect(result.current.isInvestigating).toBe(true);
     });
 
-    it('should recognize completed state without running memory', () => {
-      const { result } = renderUseInvestigation({
-        investigationPhase: InvestigationPhase.COMPLETED,
-        runningMemory: undefined,
-        historyMemory: {
-          executorMemoryId: 'mem-123',
-          parentInteractionId: 'parent-123',
-        },
+    it('should call addError with error.message as "Invalid per agent response" and i18n cause when response validation fails', async () => {
+      mockIsValidPERAgentInvestigationResponse.mockReturnValue(false);
+
+      const invalidResponse = '{"findings": [], "hypotheses": []}';
+      mockSharedMessagePollingService.poll.mockReturnValue(of(invalidResponse));
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
       });
 
-      expect(result.current.isInvestigating).toBe(false);
+      expect(mockAddError).toHaveBeenCalledWith(
+        expect.objectContaining({
+          title: 'Failed to parse response',
+          error: expect.objectContaining({
+            message: 'Invalid per agent response',
+            cause:
+              'The investigation response format is invalid. Please try running the investigation again.',
+          }),
+        })
+      );
+    });
+
+    it('should call addError with empty error.message and cause containing raw message when JSON parsing fails', async () => {
+      const invalidJsonResponse = '{"invalid": json}';
+      mockSharedMessagePollingService.poll.mockReturnValue(of(invalidJsonResponse));
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mockAddError).toHaveBeenCalledWith(
+        expect.objectContaining({
+          title: 'Failed to parse response',
+          error: expect.objectContaining({
+            message: '',
+            cause: invalidJsonResponse,
+          }),
+        })
+      );
+    });
+
+    it('should update investigation error state when response validation fails', async () => {
+      mockIsValidPERAgentInvestigationResponse.mockReturnValue(false);
+
+      const invalidResponse = '{"some": "data"}';
+      mockSharedMessagePollingService.poll.mockReturnValue(of(invalidResponse));
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mockAddError).toHaveBeenCalledWith(
+        expect.objectContaining({
+          title: 'Failed to parse response',
+          error: expect.objectContaining({
+            message: 'Invalid per agent response',
+            cause:
+              'The investigation response format is invalid. Please try running the investigation again.',
+          }),
+        })
+      );
+
+      expect(mockUpdateHypotheses).toHaveBeenCalled();
+    });
+
+    it('should handle successful response validation and parsing', async () => {
+      mockIsValidPERAgentInvestigationResponse.mockReturnValue(true);
+
+      const validResponse = JSON.stringify({
+        findings: [
+          {
+            id: 'f1',
+            description: 'Finding 1',
+            importance: 0.9,
+            evidence: 'Evidence text',
+            type: 'observation',
+          },
+        ],
+        hypotheses: [
+          {
+            id: 'h1',
+            title: 'Hypothesis 1',
+            description: 'Hypothesis description',
+            likelihood: 0.8,
+            status: 'pending',
+            supporting_findings: ['f1'],
+          },
+        ],
+        topologies: [],
+      });
+      mockSharedMessagePollingService.poll.mockReturnValue(of(validResponse));
+
+      mockParagraphHooks.batchCreateParagraphs.mockResolvedValue({
+        paragraphs: [{ id: 'paragraph-1' }],
+      });
+
+      mockParagraphHooks.batchRunParagraphs.mockResolvedValue({});
+
+      mockParagraphHooks.batchDeleteParagraphs.mockResolvedValue({});
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mockAddError).not.toHaveBeenCalled();
+
+      expect(mockUpdateHypotheses).toHaveBeenCalledWith(
+        expect.arrayContaining([
+          expect.objectContaining({
+            id: 'h1',
+            title: 'Hypothesis 1',
+            likelihood: 0.8,
+          }),
+        ]),
+        [],
+        true
+      );
+    });
+
+    it('should clean up "Max Steps Limit (xx) Reached" error message to "Max Steps Limit Reached"', async () => {
+      const maxStepsErrorMessage = 'Max Steps Limit (20) Reached';
+      mockSharedMessagePollingService.poll.mockReturnValue(of(maxStepsErrorMessage));
+
+      const { result } = renderHook(() => useInvestigation(), { wrapper });
+
+      await act(async () => {
+        await result.current.doInvestigate({
+          investigationQuestion: 'Test question',
+        });
+      });
+
+      expect(mockAddError).toHaveBeenCalledWith(
+        expect.objectContaining({
+          title: 'Failed to parse response',
+          error: expect.objectContaining({
+            message: 'Max Steps Limit Reached',
+            cause: maxStepsErrorMessage,
+          }),
+        })
+      );
     });
   });
 });
